refactor(analysis-types): split baseline DTOs
